1.1.14.0版本发布版； 2.添加优惠券，停车场导航

+++ Udive_az/app/src/main/java/cn/com/i_zj/udrive_az/model/AccountInfoResult.java
@@ -27,13 +27,13 @@ public class AccountInfoResult {
         public int depositState;
 
         @SerializedName("balance")
-        public float balance;
+        public long balance;
 
         @SerializedName("idCardState")
         public int idCardState;
 
         @SerializedName("giveBalance")
-        public int giveBalance;
+        public long giveBalance;
         @SerializedName("driverState")
         public int driverState;
     }
